move `wcsinfo_to_wcs` to `lib`, and remove unused functions
