+++ romancal/skycell/skymap.py
@@ -12,16 +12,12 @@ import spherical_geometry.great_circle_arc as sga
 import spherical_geometry.polygon as sgp
 import spherical_geometry.vector as sgv
 from asdf import AsdfFile
-from astropy import coordinates
-from astropy import units as u
-from astropy.modeling import models
-from gwcs import WCS, coordinate_frames
+from gwcs import WCS
 from numpy.typing import NDArray
-from roman_datamodels import stnode
 from scipy.spatial import KDTree
-from stcal.alignment import util as wcs_util
 
 from romancal.datamodels.library import ModelLibrary
+from romancal.lib.wcsinfo_to_wcs import wcsinfo_to_wcs
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -470,147 +466,6 @@ class ProjectionRegion:
             return False
 
         return self.data == other.data
-
-
-def wcsinfo_to_wcs(
-    wcsinfo: dict | stnode.Wcsinfo,
-    bounding_box: None | tuple[tuple[float, float], tuple[float, float]] = None,
-) -> WCS:
-    """Create a WCS from the L3 wcsinfo meta
-
-    Parameters
-    ----------
-    wcsinfo : dict or MosaicModel.meta.wcsinfo
-        The L3 wcsinfo to create a WCS from.
-
-    bounding_box : None or 4-tuple
-        The bounding box in detector/pixel space. Form of input is:
-        (x_left, x_right, y_bottom, y_top)
-
-    Returns
-    -------
-    wcs : WCS
-        The WCS object created.
-    """
-    pixelshift = models.Shift(-wcsinfo["x0_projection"], name="offset") & models.Shift(
-        -wcsinfo["y0_projection"], name="offset"
-    )
-    pixelscale = models.Scale(wcsinfo["pixel_scale"], name="scale") & models.Scale(
-        wcsinfo["pixel_scale"], name="scale"
-    )
-    tangent_projection = models.Pix2Sky_TAN()
-    celestial_rotation = models.RotateNative2Celestial(
-        wcsinfo["ra_projection_center"], wcsinfo["dec_projection_center"], 180.0
-    )
-
-    matrix = wcsinfo.get("rotation_matrix", None)
-    if matrix is not None:
-        matrix = np.array(matrix)
-    else:
-        matrix = np.reshape(
-            wcs_util.calc_rotation_matrix(
-                np.deg2rad(wcsinfo.get("orientat_projection_center", 0.0)),
-                v3i_yangle=0.0,
-                vparity=1,
-            ),
-            (2, 2),
-        )
-    rotation = models.AffineTransformation2D(matrix, name="rotation")
-    det2sky = (
-        pixelshift | rotation | pixelscale | tangent_projection | celestial_rotation
-    )
-    det2sky.name = "linear_transform"
-
-    detector_frame = coordinate_frames.Frame2D(
-        name="detector", axes_names=("x", "y"), unit=(u.pix, u.pix)
-    )
-    sky_frame = coordinate_frames.CelestialFrame(
-        reference_frame=coordinates.ICRS(), name="icrs", unit=(u.deg, u.deg)
-    )
-    wcsobj = WCS(
-        [(detector_frame, det2sky), (sky_frame, None)], name=wcsinfo.get("name", None)
-    )
-
-    if bounding_box:
-        wcsobj.bounding_box = bounding_box
-
-    return wcsobj
-
-
-def skycell_to_wcs(skycell_record: dict) -> WCS:
-    """From a skycell record, generate a WCS
-
-    Parameters
-    ----------
-    skycell_record : dict
-        A skycell record, or row, from the skycell patches table.
-
-    Returns
-    -------
-    wcsobj : WCS
-        The WCS object from the skycell record.
-    """
-    wcsinfo = dict()
-
-    # The scale is given in arcseconds per pixel. Convert to degrees.
-    wcsinfo["pixel_scale"] = float(skycell_record["pixel_scale"]) / 3600.0
-
-    # Remaining components of the wcsinfo block
-    wcsinfo["ra_ref"] = float(skycell_record["ra_projection_center"])
-    wcsinfo["dec_ref"] = float(skycell_record["dec_projection_center"])
-    wcsinfo["x_ref"] = float(skycell_record["x0_projection"])
-    wcsinfo["y_ref"] = float(skycell_record["y0_projection"])
-    wcsinfo["orientat"] = float(skycell_record["orientat_projection_center"])
-
-    # Bounding box of the skycell. Note that the center of the pixels are at (0.5, 0.5)
-    bounding_box = (
-        (-0.5, -0.5 + skycell_record["nx"]),
-        (-0.5, -0.5 + skycell_record["ny"]),
-    )
-
-    wcsobj = wcsinfo_to_wcs(wcsinfo, bounding_box=bounding_box)
-
-    wcsobj.array_shape = tuple(
-        int(axs[1] - axs[0] + 0.5)
-        for axs in wcsobj.bounding_box.bounding_box(order="C")
-    )
-    return wcsobj
-
-
-def to_skycell_wcs(library: ModelLibrary) -> WCS | None:
-    """If available read the skycell WCS from the input library association.
-
-    If the association information contains a "skycell_wcs_info" entry that
-    is not "none" it will be interpreted as a skycell wcs. If not, the
-    association "target" name will be checked. If it matches a skycell
-    name the skycell table will be loaded and a WCS constructed based on the name.
-    If neither condition is met None will be returned.
-
-    Parameters
-    ----------
-    library : ModelLibrary
-        ModelLibrary instance containing association information.
-
-    Returns
-    -------
-    wcsobj : WCS or None
-        The WCS object from the skycell record or None if
-        none was found.
-    """
-
-    try:
-        skycell = SkyCell.from_asn(library.asn)
-
-        log.info(f"Skycell record: {skycell.data}")
-
-        # extract the wcs info from the record for skycell_to_wcs
-        log.info(
-            f"Creating skycell image at ra: {skycell.radec_center[0]}  dec {skycell.radec_center[1]}",
-        )
-
-        return skycell.wcs
-    except ValueError:
-        return None
 
 
 class SkyMap:
